refactor(sidebar): move width tests to sibling module
Keep the expanded state-machine suite in its single conventional home without exceeding the inline-test size invariant.
